Short handleClick Function
