+++ site/server/src/lib/components/Exact.svelte
@@ -7,6 +7,7 @@
   import { updatePointsDisplay } from "$lib/scripts/Exact/UpdatePointsDisplay";
   import { findWinner } from "$lib/scripts/Exact/findWinner";
   import { resetGame } from "$lib/scripts/Exact/resetGame";
+  import { ifMatch } from "$lib/scripts/Exact/ifMatch";
 
   let rows = [
     {
@@ -79,62 +80,17 @@
     const match = targetId.match(/row(\d+)-(\d+)/);
 
     if (match) {
-      const rowNumber = parseInt(match[1]);
-      const index = parseInt(match[2]);
-      let clickedCell: HTMLElement | null = document.getElementById(targetId);
-      if (clickedCell) {
-        clickedCell.style.backgroundColor = color;
-        let team = teams.find((t) => t.color === color);
-        if (team) {
-          pointsByTeam[color].update((currentPoints: number) => {
-            let newPoints = currentPoints;
-
-            if (rowNumber >= 1 && rowNumber <= 6) {
-              newPoints += 1;
-            } else if (rowNumber === 7) {
-              newPoints += 2;
-            } else if (rowNumber === 8) {
-              newPoints += 3;
-            } else if (rowNumber === 9) {
-              newPoints += 5;
-            } else if (rowNumber === 10) {
-              newPoints -= 1;
-            }
-
-            if (lastRowNumbers[color] === rowNumber) {
-              if (rowNumber >= 1 && rowNumber <= 6) {
-                newPoints += 1;
-              } else if (rowNumber === 7) {
-                newPoints += 2;
-              } else if (rowNumber === 8) {
-                newPoints += 3;
-              } else if (rowNumber === 9) {
-                newPoints += 5;
-              } else if (rowNumber === 10) {
-                newPoints -= 1;
-              }
-            }
-
-            lastRowNumbers[color] = rowNumber; // Setze lastRowNumber nach der Verarbeitung
-
-            const storedData = localStorage.getItem(
-              `exact_${teams.length}_data`
-            );
-            let parsedData = storedData ? JSON.parse(storedData) : {};
-
-            parsedData[color] = { points: newPoints, shots: clickedCellsCount };
-            localStorage.setItem(
-              `exact_${teams.length}_data`,
-              JSON.stringify(parsedData)
-            ); // Im localStorage speichern
-            return newPoints;
-          });
-
-          updatePointsDisplay(teams);
-          clickedCellsCount++;
-        }
-        changeTeam(currentTeam, currentTeamIndex, color, teams);
-      }
+      ifMatch(
+        match,
+        targetId,
+        color,
+        teams,
+        pointsByTeam,
+        lastRowNumbers,
+        clickedCellsCount,
+        currentTeam,
+        currentTeamIndex
+      );
     }
 
     if (clickedCellsCount === userInput * teams.length) {
